+++ feature/clock/src/commonMain/kotlin/io/github/kabos/bus/feature/clock/ClockViewModel.kt
@@ -1,4 +1,4 @@
-package io.github.kabos
+package io.github.kabos.bus.feature.clock
 
 import androidx.compose.ui.platform.UriHandler
 import androidx.lifecycle.ViewModel
@@ -6,10 +6,6 @@ import androidx.lifecycle.viewModelScope
 import com.github.michaelbull.result.andThen
 import com.github.michaelbull.result.onFailure
 import com.github.michaelbull.result.runCatching
-import io.github.kabos.ClockContract.SideEffect
-import io.github.kabos.ClockContract.UiAction
-import io.github.kabos.ClockContract.UiState
-import io.github.kabos.ClockContract.UiState.Init
 import io.github.kabos.bus.core.domain.GetBusDepartureTimeUseCase
 import io.github.kabos.bus.core.domain.mvi.MVI
 import io.github.kabos.bus.core.domain.mvi.mviDelegate
@@ -18,6 +14,10 @@ import io.github.kabos.bus.core.model.BusRouteName
 import io.github.kabos.bus.core.model.DayType
 import io.github.kabos.bus.core.model.StationName
 import io.github.kabos.bus.core.model.TimetableCell
+import io.github.kabos.bus.feature.clock.ClockContract.SideEffect
+import io.github.kabos.bus.feature.clock.ClockContract.UiAction
+import io.github.kabos.bus.feature.clock.ClockContract.UiState
+import io.github.kabos.bus.feature.clock.ClockContract.UiState.Init
 import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
 import kotlinx.datetime.LocalTime
