+++ feature/clock/src/commonMain/kotlin/io/github/kabos/bus/feature/clock/StationSelectDialog.kt
@@ -1,4 +1,4 @@
-package io.github.kabos
+package io.github.kabos.bus.feature.clock
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
