+++ feature/clock/src/commonTest/kotlin/io/github/kabos/bus/feature/clock/TimelineSectionTest.kt
@@ -1,4 +1,4 @@
-package io.github.kabos
+package io.github.kabos.bus.feature.clock
 
 import kotlinx.datetime.LocalTime
 import kotlin.test.Test
